rename upper case files
