+++ spiketoolkit/curation/curationsortingextractor.py
